Re-structure data_source and data_sink in modules
