+++ ytui-ui/src/components/searchbar.rs
@@ -2,7 +2,7 @@ use ratatui::{
     buffer::Buffer,
     layout::Rect,
     style::{Color, Style, Stylize},
-    text::{Span, StyledGrapheme},
+    text::Span,
     widgets::{Block, BorderType, Borders, Paragraph, WidgetRef},
 };
